diving and creating mad comps
